fix crash!
application crashed as DbHelper should be extending SqlAssetsHelper NOT SqlOpenHelper, also the methods that perform queries, must check if database was opened or the connection is closed. For the channel layout, the id was made wrong. I don't know how this happens, as I pushed it with correct id. Probably Ibrahim typo it. Last thing for fragments, their lists are static so far only query database once